Refactored DisableServiceAtomicOperation.
